test - deleting a file
